Refactor widgets to load sessions directly.
